- moved Entity Info header to new component - moved Entity Info menu to new component - renamed Tombstone -> EntityDefinitions - misc cleanup - fixed a getter name - updated/created unit test suites for affected components
